requirements.txt
# Core Data Analysis Libraries
numpy==1.26.4
pandas==2.2.1
scipy==1.12.0
scikit-learn

python-dotenv

# Visualization
matplotlib==3.8.3
seaborn==0.13.2

# Claude API Interaction
anthropic

# Utility and Development
python-dotenv==1.0.1  # Environment variable management
tqdm==4.66.2          # Progress bars

# data extractor
SQLAlchemy==2.0.40
textract==1.6.3
google-api-python-client
google-auth


langchain-mcp-adapters
langgraph
langchain-anthropic
langchain-mcp
mcp==1.6.0
tavily-python